moved testhelper out of internal directory into pkg directory. with 'vendor'ing, there are golint errors with the vendored internal directory.
