+++ endpointmanager/pkg/testhelper/testclient.go
