+++ endpointmanager/pkg/testhelper/testhelpers.go
